cody/requirements.txt
ipykernel
selenium
openai
tomli
matlabengine
streamlit
pandas